[ GW ] move context creation functions from xively.c to xi_context.c/.h to make it available for internal usage in libxivelydriver (for ftests) and integration tests, and for the new GW layerchain (xi_gw.c)
