+++ src/libxively/gateway/xi_gw.c
@@ -9,17 +9,7 @@
 #include <xi_gw_layer_stack.h>
 #include <xi_handle.h>
 #include <xi_globals.h>
-
-// todo: move this function predeclaration to an internal header
-xi_state_t xi_create_context_with_custom_layers( xi_context_t** context,
-                                                 xi_layer_type_t layer_config[],
-                                                 xi_layer_type_id_t layer_chain[],
-                                                 size_t layer_chain_size );
-
-xi_state_t xi_delete_context_with_custom_layers( xi_context_t** context,
-                                                 xi_layer_type_t layer_config[],
-                                                 size_t layer_chain_size );
-
+#include <xi_context.h>
 
 xi_context_handle_t xi_create_gateway_context()
 {
