+++ src/tests/tools/xi_libxively_driver/xi_libxively_driver_impl.c
@@ -7,20 +7,9 @@
 #include "xi_driver_control_channel_layerchain.h"
 #include "xi_control_channel_protocol.pb-c.h"
 #include "xi_globals.h"
+#include <xi_context.h>
 
 xi_libxively_driver_t* libxively_driver = NULL;
-
-// todo: move this function predeclaration to an internal header
-xi_state_t
-xi_create_context_with_custom_layers_and_evtd( xi_context_t** context,
-                                               xi_layer_type_t layer_config[],
-                                               xi_layer_type_id_t layer_chain[],
-                                               size_t layer_chain_size,
-                                               xi_evtd_instance_t* event_dispatcher );
-
-xi_state_t xi_delete_context_with_custom_layers( xi_context_t** context,
-                                                 xi_layer_type_t layer_config[],
-                                                 size_t layer_chain_size );
 
 xi_state_t
 xi_driver_free_protobuf_callback( struct _XiClientFtestFw__XiClientCallback* callback );
